add home, sethome, deletehome command

+++ OneBlockMC/src/main/java/fr/cercusmc/oneblockmc/commands/players/DeleteCommand.java
@@ -12,14 +12,15 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
 import fr.cercusmc.oneblockmc.Main;
+import fr.cercusmc.oneblockmc.commands.OneBlockCommand;
 import fr.cercusmc.oneblockmc.islands.Island;
 import fr.cercusmc.oneblockmc.islands.ToolsIsland;
+import fr.cercusmc.oneblockmc.utils.Constantes;
 import fr.cercusmc.oneblockmc.utils.MessageUtil;
 import fr.cercusmc.oneblockmc.utils.SubCommand;
 
 public class DeleteCommand implements SubCommand {
 
-	private static final String MESSAGES = "messages";
 
 	@Override
 	public String getName() {
@@ -34,28 +35,31 @@ public class DeleteCommand implements SubCommand {
 
 	@Override
 	public String getDescription() {
-		return Main.getFiles().get(MESSAGES).getString("commands.delete.description");
+		return Main.getFiles().get(Constantes.MESSAGES).getString("commands.delete.description");
 	}
 
 	@Override
 	public String getSyntax() {
-		return Main.getFiles().get(MESSAGES).getString("commands.delete.syntax");
+		return Main.getFiles().get(Constantes.MESSAGES).getString("commands.delete.syntax");
 	}
 
 	@Override
 	public void perform(CommandSender sender, String[] args) {
+		
 		Player p = (Player) sender;
+		if(OneBlockCommand.messageTooManyArgs(args, p, this)) return;
+		
 		Island is = ToolsIsland.getIslandOfPlayer(p.getUniqueId());
 		if(is == null) {
-			MessageUtil.sendMessage(p, Main.getFiles().get(MESSAGES).getString("island.not_island"));
+			MessageUtil.sendMessage(p, Main.getFiles().get(Constantes.MESSAGES).getString("island.not_island"));
 			
 		} else {
 			if(is.playerIsOwner(p.getUniqueId())) {
 				ToolsIsland.deleteIsland(is, p.getUniqueId());
 				for(String member : is.getMembers()) {
-					if(Bukkit.getPlayer(UUID.fromString(member)).isOnline()) {
+					if(Bukkit.getPlayer(UUID.fromString(member)).isOnline() && !p.getUniqueId().toString().equals(member)) {
 						Bukkit.getPlayer(UUID.fromString(member)).teleport(Main.getIslandConfig().getSpawn());
-						MessageUtil.sendMessage(Bukkit.getPlayer(UUID.fromString(member)), Main.getFiles().get(MESSAGES).getString("island.message_member_deleted_island"));
+						MessageUtil.sendMessage(Bukkit.getPlayer(UUID.fromString(member)), Main.getFiles().get(Constantes.MESSAGES).getString("island.message_member_deleted_island"));
 					} else {
 						Set<String> players = Main.getFiles().get("deletePlayerWaiting").getStringList("players").stream().collect(Collectors.toSet());
 						players.add(member);
@@ -63,7 +67,7 @@ public class DeleteCommand implements SubCommand {
 					}
 				}
 			} else {
-				MessageUtil.sendMessage(p, Main.getFiles().get(MESSAGES).getString("island.not_owner"));
+				MessageUtil.sendMessage(p, Main.getFiles().get(Constantes.MESSAGES).getString("island.not_owner"));
 			}
 		}
